refactor: break down tools/memory.py into modular components
Massive reduction: 973 → 47 lines (95% reduction) by extracting:
- memory_modules/utils.py - JSON encoding and utilities (89 lines)
- memory_modules/search_operations.py - Search functionality (296 lines)
- memory_modules/crud_operations.py - Memory CRUD operations (350 lines)

Maintains complete backward compatibility while dramatically improving
maintainability, testability, and code organization. All functions
properly modularized with consistent error handling.
